requirements.txt
PyEnchant==1.6.6